moved types to the types file
