--- src/internal/util/ValueGetter.ts
@@ -1,3 +1,0 @@
-export interface ValueGetter<T> {
-  getValue(): T;
-}
